+++ CHANGELOG.md
@@ -1,3 +1,23 @@
+## [1.19.1-prerelease.2](https://github.com/adobe/rum-distiller/compare/v1.19.1-prerelease.1...v1.19.1-prerelease.2) (2025-10-07)
+
+
+### Bug Fixes
+
+* cache facet values in valuesExtractorFn to eliminate redundant calls ([7057860](https://github.com/adobe/rum-distiller/commit/7057860e1e219fa2fde4c3fd0e30b17c73206187)), closes [#1](https://github.com/adobe/rum-distiller/issues/1)
+
+## [1.19.1-prerelease.1](https://github.com/adobe/rum-distiller/compare/v1.19.0...v1.19.1-prerelease.1) (2025-10-06)
+
+
+### Bug Fixes
+
+* **docs:** better jsdocs ([958a513](https://github.com/adobe/rum-distiller/commit/958a5136f493b8f67c5bead1aa0de89e140837a0))
+* **docs:** typo ([9123c61](https://github.com/adobe/rum-distiller/commit/9123c6188311815465c68a783f303c207a7d41e8))
+
+
+### Performance Improvements
+
+* optimize grouping with pre-allocated arrays ([32d58e6](https://github.com/adobe/rum-distiller/commit/32d58e6af7b1b93d3ba7987a9ca3ec599d4d8f1a)), closes [#4](https://github.com/adobe/rum-distiller/issues/4)
+
 # [1.19.0](https://github.com/adobe/rum-distiller/compare/v1.18.0...v1.19.0) (2025-09-12)